--- html/options.html
@@ -1,111 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Vet Radar Notification - Settings</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            padding: 20px;
-            background-color: #f4f4f4;
-        }
-
-        h1 {
-            font-size: 24px;
-            margin-bottom: 20px;
-        }
-
-        form {
-            display: flex;
-            flex-direction: column;
-        }
-
-        label {
-            margin-bottom: 5px;
-            font-weight: bold;
-        }
-
-        input[type="file"] {
-            margin-bottom: 20px;
-        }
-
-        .button-group {
-            display: flex;
-            gap: 10px;
-        }
-
-        button {
-            padding: 10px 20px;
-            background-color: #4CAF50;
-            color: white;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-            flex: 1;
-        }
-
-        button:disabled {
-            background-color: #ccc;
-            cursor: not-allowed;
-        }
-
-        button:hover:not(:disabled) {
-            background-color: #45a049;
-        }
-
-        #disableButton {
-            background-color: #f44336;
-        }
-
-        #disableButton:hover:not(:disabled) {
-            background-color: #d7352b;
-        }
-
-        #currentSound {
-            margin-bottom: 20px;
-            font-size: 14px;
-            color: #333;
-        }
-
-        #message {
-            margin-top: 15px;
-            font-size: 14px;
-        }
-
-        #message.success {
-            color: green;
-        }
-
-        #message.error {
-            color: red;
-        }
-    </style>
-</head>
-<body>
-    <h1>Vet Radar Notification - Settings</h1>
-
-    <div id="currentSound">Current Sound: Default</div>
-	<div id="message" role="alert" aria-live="assertive"></div>
-	
-    <form id="settingsForm">
-		<label for="chimeVolume">Chime Volume:</label>
-		<input type="range" id="chimeVolume" min="0" max="1" step="0.1">
-		<span id="volumeValue">0.5</span>
-		<label for="soundPicker">Choose a sound file (MP3, max 5MB):</label>
-		<input type="file" id="soundPicker" accept="audio/mp3">
-		<audio id="soundPreview" controls style="display:none; margin-top: 20px;"></audio>
-
-
-        <div class="button-group">
-            <button type="button" id="saveSettings" disabled>Save and Close</button>
-            <button type="button" id="disableButton">Disable Extension</button>
-			<button type="button" id="resetSound">Reset to Default</button>
-        </div>
-
-        <div id="message" role="alert" aria-live="assertive"></div>
-    </form>
-
-    <script src="options.js"></script>
-</body>
-</html>
