--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "chrome-extension-patient-management-system",
-  "version": "1.0.0",
-  "main": "index.js",
-  "directories": {
-    "doc": "doc"
-  },
-  "scripts": {
-    "build": "webpack --mode production --config webpack.config.js",
-    "build:dev": "webpack --mode development --config webpack.config.js",
-    "watch": "webpack --watch --mode development --config webpack.config.js"
-  },
-  "keywords": [],
-  "author": "Tom Hoffpauir",
-  "license": "MIT",
-  "description": "Chrome extension for Vet Radar notifications",
-  "devDependencies": {
-    "@babel/core": "^7.25.8",
-    "@babel/preset-env": "^7.25.8",
-    "babel-loader": "^9.2.1",
-    "copy-webpack-plugin": "^12.0.2",
-    "webpack": "^5.95.0",
-    "webpack-cli": "^5.1.4"
-  }
-}
